dev/NS-bytesize/requirements.txt
aiohttp>=3.8.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-timeout>=2.1.0
autogen>=0.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5
pytest-cov>=2.12.0
mock>=4.0.0
typing-extensions>=4.0.0
psutil>=5.9.0
